+++ CHANGELOG.md
@@ -2,9 +2,8 @@
 
 All notable changes to the "cyezoi" extension will be documented in this file.
 
-## v0.0.19
+## v0.0.20
 
-- **Automatically refresh the record tree view when the record is updated**
-- Fix attend problem
+- Go to contest problem when viewing contest record
 
 Want to see the previous log? [Click here](https://github.com/CYEZOI/cyezoi-helper/commits/main/CHANGELOG.md)